+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## 2025.08.16  
 `+` Addition: client에서 키보드 입력 이벤트 발생 시 WebSocket으로 연결된 서버로 보낼 수 있도록 기능 구현 중...
+`-` Fix: 서버에서 `test.rs`에서 테스트 클라이언트로 방향키를 눌렀을 때 받는 코드 작성 완료.  
     
 # 0.1.0
 `@` Version: first commit